requirements.txt
discord.py == 0.16.12
PyGithub
pyssml
gspread
oauth2client
PyOpenSSL
